+++ app/common/build.gradle
@@ -1,0 +1,28 @@
+buildscript {
+    apply from: "$rootProject.projectDir/gradle/buildscript.gradle", to: buildscript
+}
+
+description = "Example - Common"
+defaultTasks = ['commonDeploy']
+
+apply plugin: 'java'
+apply plugin: "kotlin"
+apply plugin: 'org.dm.bundle'
+apply plugin: 'cognifide.aem'
+
+bundle {
+    def pkg = 'com.company.aem.example.common'
+
+    instruction 'Bundle-SymbolicName', pkg
+    instruction 'Sling-Model-Packages', pkg
+    instruction 'Export-Package', "$pkg.*;-split-package:=merge-first,org.hashids.*"
+}
+
+dependencies {
+    aemInstall group: 'org.jetbrains.kotlin', name: 'kotlin-osgi-bundle', version: '1.1.1'
+    aemInstall group: 'org.jsoup', name: 'jsoup', version: '1.10.2'
+    aemEmbed group: 'org.hashids', name: 'hashids', version: '1.0.1'
+}
+
+build.dependsOn aemCompose
+task commonDeploy(dependsOn: [build, aemDeploy])
