+++ app/common/src/main/kotlin/com/company/aem/example/common/AppUtils.kt
