+++ app/core/build.gradle
@@ -1,0 +1,26 @@
+buildscript {
+    apply from: "$rootProject.projectDir/gradle/buildscript.gradle", to: buildscript
+}
+
+description = "Example - Core"
+defaultTasks = ['coreDeploy']
+
+apply plugin: 'java'
+apply plugin: "kotlin"
+apply plugin: 'org.dm.bundle'
+apply plugin: 'cognifide.aem'
+
+bundle {
+    def pkg = 'com.company.aem.example.core'
+
+    instruction 'Bundle-SymbolicName', pkg
+    instruction 'Sling-Model-Packages', pkg
+    instruction 'Export-Package', "$pkg.*"
+}
+
+dependencies {
+    compile project(':app:common')
+}
+
+build.dependsOn aemCompose
+task coreDeploy(dependsOn: [build, aemDeploy])
