+++ app/core/src/main/content/jcr_root/apps/example/core/install/.vltignore
@@ -1,0 +1,1 @@
+*.jar
