+++ app/core/src/main/kotlin/com/company/aem/example/core/services/posts/Post.kt
@@ -1,0 +1,8 @@
+package com.company.aem.example.core.services.posts
+
+data class Post(
+    val userId : Int,
+    val id : Int,
+    val title: String,
+    val body : String
+)
