+++ app/core/src/test/java/com/company/aem/example/core/SimpleUnitTest.java
@@ -1,0 +1,14 @@
+package com.company.aem.example.core;
+
+import static org.junit.Assert.*;
+
+import org.junit.Test;
+
+public class SimpleUnitTest {
+
+    @Test
+    public void someTest() {
+        assertTrue(true);
+    }
+
+}
