+++ app/design/src/main/content/jcr_root/etc/designs/example/author/css.txt
@@ -1,0 +1,2 @@
+#base=css
+main.less
