+++ app/design/src/main/content/jcr_root/etc/designs/example/author/css/main..less
@@ -1,0 +1,1 @@
+/* Example Author LESS */
