+++ app/design/src/main/content/jcr_root/etc/designs/example/author/js.txt
@@ -1,0 +1,2 @@
+#base=js
+main.js
