+++ app/design/src/main/content/jcr_root/etc/designs/example/publish/css/main..less
@@ -1,0 +1,11 @@
+body {
+  background-color: #ddd;
+}
+
+a {
+  text-decoration: none;
+
+  &:hover {
+    text-decoration: underline;
+  }
+}
