+++ app/design/src/main/content/jcr_root/etc/designs/example/publish/js.txt
@@ -1,0 +1,2 @@
+#base=js
+main.js
