+++ app/design/src/main/content/jcr_root/etc/designs/example/publish/js/main.js
@@ -1,0 +1,3 @@
+(function () {
+    console.log('Example Publish JS')
+}());
