+++ app/src/main/content/META-INF/vault/config.xml
