+++ app/src/main/content/META-INF/vault/filter.xml
@@ -1,0 +1,4 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<workspaceFilter version="1.0">
+    ${includedFilterRoots}
+</workspaceFilter>
