--- bundle/build.gradle
@@ -1,26 +1,0 @@
-buildscript {
-    apply from: "$rootProject.projectDir/gradle/buildscript.gradle", to: buildscript
-}
-
-description = "Example - Bundle"
-defaultTasks = ['bundleDeploy']
-
-apply plugin: 'java'
-apply plugin: "kotlin"
-apply plugin: 'org.dm.bundle'
-apply plugin: 'cognifide.aem'
-
-bundle {
-    instruction 'Bundle-SymbolicName', 'com.example.aem.bundle'
-    instruction 'Sling-Model-Packages', 'com.example.aem.bundle'
-    instruction 'Export-Package', 'com.example.aem.bundle.*;-split-package:=merge-first,org.hashids.*'
-}
-
-dependencies {
-    aemInstall group: 'org.jetbrains.kotlin', name: 'kotlin-osgi-bundle', version: '1.1.1'
-    aemInstall group: 'org.jsoup', name: 'jsoup', version: '1.10.2'
-    aemEmbed group: 'org.hashids', name: 'hashids', version: '1.0.1'
-}
-
-build.dependsOn aemCompose
-task bundleDeploy(dependsOn: [build, aemDeploy])
