--- bundle/src/main/kotlin/com/example/aem/bundle/services/posts/Post.kt
@@ -1,8 +1,0 @@
-package com.example.aem.bundle.services.posts
-
-data class Post(
-    val userId : Int,
-    val id : Int,
-    val title: String,
-    val body : String
-)
