--- bundle/src/test/java/com/example/aem/bundle/SimpleUnitTest.java
@@ -1,14 +1,0 @@
-package com.example.aem.bundle;
-
-import static org.junit.Assert.*;
-
-import org.junit.Test;
-
-public class SimpleUnitTest {
-
-    @Test
-    public void someTest() {
-        assertTrue(true);
-    }
-
-}
