--- content/build.gradle
@@ -1,32 +1,0 @@
-buildscript {
-    apply from: "$rootProject.projectDir/gradle/buildscript.gradle", to: buildscript
-}
-
-description = "Example - Content"
-defaultTasks = ['contentDeploy']
-
-apply plugin: 'cognifide.aem'
-
-aem {
-    config {
-        contentPath = "src/main/content"
-    }
-}
-
-aemSatisfy {
-    group('dependencies', {
-        // local("pkg/vanityurls-components-1.0.2.zip")
-    })
-
-    group('tools', {
-        download('https://github.com/OlsonDigital/aem-groovy-console/releases/download/9.0.1/aem-groovy-console-9.0.1.zip')
-        download("https://github.com/Cognifide/APM/releases/download/cqsm-2.0.0/apm-2.0.0.zip")
-    })
-}
-
-aemCompose {
-    includeBundles ':example.bundle'
-}
-
-build.dependsOn aemCompose
-task contentDeploy(dependsOn: [build, aemDeploy])
