--- content/src/main/content/META-INF/vault/filter.xml
@@ -1,4 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<workspaceFilter version="1.0">
-    <filter root="/apps/example"/>
-</workspaceFilter>
