+++ gradle/buildscript.gradle
@@ -23,7 +23,7 @@ repositories {
 }
 
 dependencies {
-    classpath 'com.cognifide.gradle:aem-plugin:1.1.+'
+    classpath 'com.cognifide.gradle:aem-plugin:1.1.2-SNAPSHOT'
     classpath "org.dm.gradle:gradle-bundle-plugin:0.10.0"
     classpath "org.jetbrains.kotlin:kotlin-gradle-plugin:1.1.1"
 }
