+++ settings.gradle
@@ -1,7 +1,7 @@
 rootProject.name = 'example'
 
-include ':example.bundle'
-project(':example.bundle').projectDir = "$rootDir/bundle" as File
-
-include ':example.content'
-project(':example.content').projectDir = "$rootDir/content" as File
+include ':app'
+include ':app:common'
+include ':app:core'
+include ':app:config'
+include ':app:design'
